+++ packages/sources/1forge/test/integration/fixtures.ts
@@ -1,6 +1,6 @@
 import nock from 'nock'
 
-export const mockResponseSuccess = (): nock =>
+export const mockResponseSuccessConvertEndpoint = (): nock =>
   nock('https://api.1forge.com', {
     encodedQueryParams: true,
   })
@@ -13,6 +13,35 @@ export const mockResponseSuccess = (): nock =>
         text: '1 USD is worth 0.862701 EUR',
         timestamp: 1636478097478,
       }),
+      [
+        'Content-Type',
+        'application/json',
+        'Connection',
+        'close',
+        'Vary',
+        'Accept-Encoding',
+        'Vary',
+        'Origin',
+      ],
+    )
+
+export const mockResponseSuccessQuotesEnpoint = (): nock =>
+  nock('https://api.1forge.com', {
+    encodedQueryParams: true,
+  })
+    .get('/quotes')
+    .query({ api_key: 'fake-api-key', pairs: 'USD/EUR' })
+    .reply(
+      200,
+      (_, request) => [
+        {
+          p: 0.8828,
+          a: 0.8828,
+          b: 0.8827,
+          s: 'USD/EUR',
+          t: 1641851954307,
+        },
+      ],
       [
         'Content-Type',
         'application/json',