+++ packages/sources/cryptoapis/src/endpoint/bc_info.ts
@@ -9,9 +9,34 @@ export const endpointResultPaths = {
 }
 
 export const inputParameters: InputParameters = {
-  blockchain: ['blockchain', 'coin'],
-  resultPath: false,
-  network: false,
+  blockchain: {
+    aliases: ['coin', 'market'],
+    description: 'The blockchain to retrieve info for',
+    options: ['BTC', 'ETH', 'LTC', 'ETC', 'BCH', 'DOGE', 'DASH'],
+    required: true,
+    type: 'string',
+  },
+  network: {
+    description: 'The blockchain network name',
+    default: 'mainnet',
+    required: true,
+    type: 'string',
+  },
+}
+
+export interface ResponseSchema {
+  payload: {
+    difficulty: number
+    headers: number
+    chain: string
+    chainWork: string
+    mediantime: number
+    blocks: number
+    bestBlockHash: string
+    currency: string
+    transactions: number
+    verificationProgress: number
+  }
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
@@ -26,8 +51,8 @@ export const execute: ExecuteWithConfig<Config> = async (request, _, config) => 
 
   const reqConfig = { ...config.api, url }
 
-  const response = await Requester.request(reqConfig)
-  response.data.result = Requester.validateResultNumber(response.data, ['payload', resultPath])
+  const response = await Requester.request<ResponseSchema>(reqConfig)
+  const result = Requester.validateResultNumber(response.data, ['payload', resultPath])
 
-  return Requester.success(jobRunID, response, config.verbose)
+  return Requester.success(jobRunID, Requester.withResult(response, result), config.verbose)
 }