+++ packages/sources/dxfeed/src/endpoint/price.ts
@@ -14,7 +14,11 @@ export const batchablePropertyPath = [{ name: 'base', limit: 120 }]
 const customError = (data: { status: string }) => data.status !== 'OK'
 
 export const inputParameters: InputParameters = {
-  base: ['base', 'from', 'coin', 'market'],
+  base: {
+    required: true,
+    aliases: ['from', 'coin', 'market'],
+    description: 'The symbol of the currency to query',
+  },
 }
 
 const quoteEventSymbols: { [key: string]: boolean } = {