+++ packages/sources/flightaware/src/endpoint/flightinfoex.ts
@@ -39,9 +39,16 @@ export interface ResponseSchema {
 }
 
 export const inputParameters: InputParameters = {
-  departure: true,
-  flight: true,
-  resultPath: false,
+  departure: {
+    required: true,
+    description: 'The departure time of the flight as a UNIX timestamp in seconds',
+    type: 'number',
+  },
+  flight: {
+    required: true,
+    description: 'The flight ID',
+    type: 'string',
+  },
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {