+++ packages/sources/geodb/src/endpoint/matches.ts
@@ -4,11 +4,35 @@ import { ExecuteWithConfig, Config, InputParameters } from '@chainlink/types'
 export const supportedEndpoints = ['matches']
 
 export const inputParameters: InputParameters = {
-  lat: true,
-  lng: true,
-  radius: true,
-  start: true,
-  end: true,
+  lat: {
+    required: true,
+    description: 'latitude coordinate',
+    type: 'string',
+  },
+  lng: {
+    required: true,
+    description: 'longitude coordinate',
+    type: 'string',
+  },
+  radius: {
+    required: true,
+    description: 'radius around coordinates (in m)',
+    type: 'string',
+  },
+  start: {
+    required: true,
+    description: 'start time (yyyy-mm-dd hh:mm:ss)',
+    type: 'string',
+  },
+  end: {
+    required: true,
+    description: 'end time (yyyy-mm-dd hh:mm:ss)',
+    type: 'string',
+  },
+}
+
+export interface ResponseSchema {
+  matches: string
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
@@ -28,8 +52,8 @@ export const execute: ExecuteWithConfig<Config> = async (request, _, config) => 
     url,
   }
 
-  const response = await Requester.request(reqConfig)
-  response.data.result = parseInt(response.data.matches)
+  const response = await Requester.request<ResponseSchema>(reqConfig)
+  const result = Requester.validateResultNumber(response.data, ['matches'])
 
-  return Requester.success(jobRunID, response, config.verbose)
+  return Requester.success(jobRunID, Requester.withResult(response, result), config.verbose)
 }