+++ packages/sources/iex-cloud/src/endpoint/crypto.ts
@@ -5,8 +5,32 @@ import { NAME as AdapterName } from '../config'
 export const supportedEndpoints = ['crypto']
 
 export const inputParameters: InputParameters = {
-  base: ['base', 'from', 'coin', 'asset', 'symbol'],
-  quote: ['quote', 'to', 'market'],
+  base: {
+    aliases: ['from', 'coin', 'asset', 'symbol'],
+    description: 'The symbol to query',
+    required: true,
+    type: 'string',
+  },
+  quote: {
+    aliases: ['to', 'market'],
+    description: 'The symbol to convert to',
+    required: true,
+    type: 'string',
+  },
+}
+
+export interface ResponseSchema {
+  symbol: string
+  primaryExchange: string
+  sector: string
+  calculationPrice: string
+  high: string
+  low: string
+  latestPrice: string
+  latestSource: string
+  latestUpdate: number
+  latestVolume: string
+  previousClose: string
 }
 
 export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
@@ -28,8 +52,8 @@ export const execute: ExecuteWithConfig<Config> = async (request, _, config) => 
     url,
   }
 
-  const response = await Requester.request(reqConfig)
-  response.data.result = Requester.validateResultNumber(response.data, ['latestPrice'])
+  const response = await Requester.request<ResponseSchema>(reqConfig)
+  const result = Requester.validateResultNumber(response.data, ['latestPrice'])
 
-  return Requester.success(jobRunID, response, config.verbose)
+  return Requester.success(jobRunID, Requester.withResult(response, result), config.verbose)
 }